+++ lib/api/dt.dart
@@ -14,7 +14,7 @@ class DT {
   /// - [channel] 渠道，默认为空字符串，可用 ROIQueryChannel.GP，具体联系商务
   /// - [isDebug] 是否打开调试，调试模式下将打印log,默认关闭
   /// - [logLevel] log 的级别，默认为 Log.VERBOSE，仅在 isDebug = true 有效
-  static void initSDK(
+  static Future<void> initSDK(
       String appId,
       String url,
       String channel,
@@ -26,9 +26,12 @@ class DT {
       typeSuffix = "-iOS";
     } else if (Platform.isAndroid) {
       typeSuffix = "-Android";
+    } else {
+      // Unsupported platforms
+      typeSuffix = "-${Platform.operatingSystem}";
     }
 
-    DTPigeon().initSDK(appId, url, channel, isDebug, logLevel, {
+    return DTPigeon().initSDK(appId, url, channel, isDebug, logLevel, {
       "#sdk_type": "Flutter$typeSuffix",
       "#sdk_version_name": _versionName,
     });