+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Added in-code doc for all APIs.
 * Upgraded DT Android to 2.0.6.
 * Restrict DTAnalytics.userAppend, DTAnalytics.userUniqAppend to only accept List as value in JsonMap.  
+* Changes return type for all of functions to Future, to allows error handling and run blocking.  
 
 # 2.0.0